pyproject.toml
[project]
name = "energy-price-forecast"
version = "0.1.0"
description = "Day-ahead electricity price forecasting for Germany (DE-LU) with LightGBM, quantile regression, and walk-forward backtesting."
readme = "README.md"
authors = [
    { name = "Soentke Blindow", email = "270722541+soentkeblindow@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "entsoe-py>=0.8.0",
    "holidays>=0.95",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.11.1",
    "optuna>=4.8.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "statsmodels>=0.14.6",
    "streamlit>=1.56.0",
    "yfinance>=1.3.0",
]

[project.scripts]
energy-price-forecast = "energy_price_forecast:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "freezegun>=1.5.0",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "mypy>=1.20.2",
    "pandas-stubs>=3.0.0.260204",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
    "types-python-dateutil>=2.9.0.20260408",
    "types-requests>=2.32.0",
]


# ==========================================
# Tool configurations
# ==========================================

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["notebooks", ".venv"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (modernize syntax)
    "B",    # flake8-bugbear (common bugs)
    "SIM",  # flake8-simplify (code simplification)
    "N",    # pep8-naming (naming conventions)
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
strict_optional = true
no_implicit_optional = true
# Wir starten mit moderater Strenge;
# "strict = true" können wir später anknipsen
exclude = ["notebooks/", ".venv/"]

[[tool.mypy.overrides]]
# Externe Libraries ohne Typ-Stubs — keine Fehler für deren Imports
module = [
    "entsoe.*",
    "lightgbm.*",
    "yfinance.*",
    "shap.*",
    "statsmodels.*",
    "mlflow.*",
    "streamlit.*",
    "holidays.*",
    "optuna.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests that make real ENTSO-E API calls (deselect with '-m not integration')",
]
addopts = [
    "-v",
    "--strict-markers",
    "-m", "not integration",
]
